simplify builder by moving extract_time_shifts to time_shift_state
