+++ rust/cubesqlplanner/cubesqlplanner/src/planner/planners/multi_stage/mod.rs
@@ -3,9 +3,11 @@ mod member;
 mod member_query_planner;
 mod multi_stage_query_planner;
 mod query_description;
+mod time_shift_state;
 
 pub use applied_state::*;
 pub use member::*;
 pub use member_query_planner::MultiStageMemberQueryPlanner;
 pub use multi_stage_query_planner::MultiStageQueryPlanner;
 pub use query_description::MultiStageQueryDescription;
+pub use time_shift_state::TimeShiftState;
